pyproject.toml
[tool.poetry]
name = "dimfred"
version = "0.18.0"
description = "Tools and abbrevations I use often."
authors = ["Dmitrij Vinokour <dimfred.1337@web.de>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
tabulate = "^0.8.9"
easydict = "^1.9"
PyYAML = "^6.0"
pluck = "^0.2"
urlpath = "^1.2.0"
prettyprint = "^0.1.5"
psutil = "^5.9.0"
click = "<8.1.0"
shutils = "^0.1.0"
pydantic = "^1.9.1"
jsoncomment = "^0.4.2"
python-dotenv = "^0.20.0"
PyRSMQ = "^0.4.5"

[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.18.3"
pytest-mock = "^3.7.0"
pytest = "^7.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
